+++ NEWS.md
@@ -1,5 +1,9 @@
 # ospsuite.utils 1.0.0.9000
 
+MINOR CHANGES
+
+* `validateIsCharacter()` is added as an alias for `validateIsString()` function.
+
 # ospsuite.utils 1.0.0
 
 * Initial release.